refactor: organize types tests into read/ subdirectory
gassma-test の test/<category>/ 構造に倣い、include/omit の .test-d.ts を
read/ サブディレクトリに移動。将来 write/、aggregate/ などのカテゴリを
追加するための土台。vitest typecheck.include は ** で再帰するので
パターン変更は不要。
