+++ src/__test__/types/read/omit.test-d.ts
@@ -1,5 +1,5 @@
 import { expectTypeOf } from "vitest";
-import type { GassmaClient } from "./__generated__/client";
+import type { GassmaClient } from "../__generated__/client";
 
 declare const client: GassmaClient;
 declare const clientOmit: GassmaClient<{ User: { email: true } }>;
